Moved tests to spek
